+++ CHANGELOG.md
@@ -152,6 +152,20 @@
 - Does not allow deletion of groups associated with departments
 - Fixes the bug: duplicating an AI Agent does not work
 - Block file uploads in conversation details based on permission “Able to send message and upload file”
+### 2.7.138
+- Implements caching strategy for projects, project users, and departments to reduce redundant HTTP requests and improve performance
+- Adds cache invalidation on data updates to ensure data consistency
+
+### 2.7.137
+- Adds MCP server to integrations
+
+### 2.7.136
+- Adds a link to the documentation on the New Broadcast page
+
+### 2.7.135
+- Adds the ability to select between multiple LLMs in the AI ​​settings dialog
+
+### 2.7.134
 - Adds a tooltip to the file upload button in the conversation detail that indicates the maximum size allowed
 - Displays a confirm dialog on delete Canned response 
 - Manage the Knowledge preview error “allowed max tokens”
@@ -198,6 +212,11 @@
 
 ## 2.7.132-stage
 - Adds vLLM integration
+### 2.7.133
+- Updates markdown library
+
+### 2.7.132
+- Fixes the bug: Ai Agent duplicate not working
 
 ### 2.7.131
 - Adds the ability to manage which pages the widget can appear on in Widget Settings